move CreateTarFile & CreateAptTarFile to helpers
The monobase helpers from command.Command were calling docker run, instead have them call the run functions on the client since the logic doesn't depend on client implementation
