+++ pkg/docker/command/command.go
@@ -15,8 +15,6 @@ type Command interface {
 	Pull(ctx context.Context, ref string, force bool) (*image.InspectResponse, error)
 	Push(ctx context.Context, ref string) error
 	LoadUserInformation(ctx context.Context, registryHost string) (*UserInfo, error)
-	CreateTarFile(ctx context.Context, ref string, tmpDir string, tarFile string, folder string) (string, error)
-	CreateAptTarFile(ctx context.Context, tmpDir string, aptTarFile string, packages ...string) (string, error)
 	Inspect(ctx context.Context, ref string) (*image.InspectResponse, error)
 	ImageExists(ctx context.Context, ref string) (bool, error)
 	ContainerLogs(ctx context.Context, containerID string, w io.Writer) error
